refactor(node): node as parent template class, drive_node sets folder_mimetype
- node logic now located in Node instead of DriveNode
- DriveNode converted to subclass of Node
- is_folder() checks using folder_mimetype (set in drive_node)
